Protracted Birth-Death Model of Diversification

This package computes the (maximum) likelihood of the protracted speciation model for a given set of branching times This package is a likelihood-based statistical package to estimate parameters under the protracted speciation model.

First version: 0.8
New in version 0.9
- Bug fix for stem age
New in version 0.91
- Reports loglik = -Inf on an error in the deSolve package (function ode)
New in version 0.92
- Correcting order of parameters of pbd_sim
New in version 0.93
- pbd_sim produces a tree, a matrix containing all events in the simulation, and a tree with one sample per species.
New in version 1.0
- Conditioning is also possible on a range of values of the number of species.
New in version 1.1
- Simulation of the protracted speciation tree has more features.
New in version 1.2
- Optimization can make use of subplex (default) and simplex (older versions).
New in version 1.3
- Contains a function to carry out a bootstrap likelihood ratio test.
- Vignette and test added.
- Reports an error if exteq = TRUE and initparsopt contains 4 parameters.
- Option to limit a simulation to a certain maximum number of species; if exceeded, the simulation is ignored.
New in version 1.4:
- Includes all special cases in pbd_durspec_mean
- Fixes a bug in conditioning on a range of values of the number of species
New in version 1.5:
- Fixes bug with missing species in dd_LR

Details

pbd_loglik computes the likelihood of the protracted birth-death model of diversification, given a set of parameters and a data set of phylogenetic branching times.

pbd_ML finds the parameters that maximizes the likelihood computed by pbd_loglik.

pbd_bootstrap performs a maximum likelihood analysis and simulates with the maximum likelihood parameters. The ML parameters of the simulated data sets are then estimated, providing an uncertainty distribution for the original ML estimate on the original data.

References

- Etienne, R.S. & J. Rosindell 2012. Systematic Biology 61: 204-213.
- Lambert, A., H. Morlon & R.S. Etienne 2014. Journal of Mathmematical Biology 70: 367-397. doi:10.1007/s00285-014-0767-x
- Etienne, R.S., H. Morlon & A. Lambert 2014. Evolution 68: 2430-2440, doi:10.1111/evo.12433
. - Etienne, R.S., A.L. Pigot & A.B. Phillimore 2016. Methods in Ecology & Evolution 7: 1092-1099, doi: 10.1111/2041-210X.12565

See also

Author

Rampal S. Etienne Maintainer: Rampal S. Etienne <r.s.etienne@rug.nl>