mode of the duration of speciation under protracted birth-death model of diversification

pbd_durspec_mode computes the mode of the duration of speciation under the protracted speciation model for a given set of parameters

pbd_durspec_mode(pars)

Arguments

pars

Vector of parameters:

pars[1] corresponds to b (= la_3 in Etienne & Rosindell R2012) = speciation initiation rate
pars[2] corresponds to la_1 (= la_2 in Etienne & Rosindell 2012) = speciation completion rate
pars[3] corresponds to mu_2 (= mu_i in ER2012) = extinction rate of incipient species

Value

The expected duration of speciation

See also

Author

Rampal S. Etienne

Examples

 pbd_durspec_mode(pars = c(0.5,0.3,0.1))
#> [1] 0.3296783