Function to simulate the protracted speciation process

Simulating the protracted speciation process using the Doob-Gillespie algorithm. This function differs from pbd_sim_cpp that 1) it does not require that the speciation-initiation rate is the same for good and incipient species, and 2) that it simulates the exact protracted speciation process, and not the approximation made by the coalescent point process. This function provides also the conversion to the approximation as output.

pbd_sim(
  pars,
  age,
  soc = 2,
  plotit = FALSE,
  limitsize = Inf,
  add_shortest_and_longest = FALSE
)

Arguments

pars

Vector of parameters:

pars[1] corresponds to b_1, the speciation-initiation rate of good species
pars[2] corresponds to la_1, the speciation-completion rate
pars[3] corresponds to b_2, the speciation-initiation rate of incipient species
pars[4] corresponds to mu_1, the extinction rate of good species
pars[5] corresponds to mu_2, the extinction rate of incipient species

age

Sets the age for the simulation

soc

Sets whether this age is the stem (1) or crown (2) age

plotit

Sets whether the various trees produced by the function should be plotted or not

limitsize

Sets a maximum to the number of incipient + good species that are created during the simulation; if exceeded, the simulation is aborted and removed.

add_shortest_and_longest

Gives the output of the new samplemethods 'shortest' and 'longest'.

Value

out

A list with the following elements:

tree is the tree of extant species in phylo format
stree_random is a tree with one random sample per species in phylo format
stree_oldest is a tree with the oldest sample per species in phylo format
stree_youngest is a tree with the youngest sample per species in phylo format
L is a matrix of all events in the simulation where
- the first column is the incipient-level label of a species
- the second column is the incipient-level label of the parent of the species
- the third column is the time at which a species is born as incipient species
- the fourth column is the time of speciation-completion of the species
If the fourth element equals -1, then the species is still incipient. - the fifth column is the time of extinction of the species
If the fifth element equals -1, then the species is still extant. - The sixth column is the species-level label of the species
sL_random is a matrix like L but for stree_random
sL_oldest is a matrix like L but for stree_oldest
sL_youngest is a matrix like L but for stree_youngest
igtree.extinct is the tree in simmap format with incipient and good flags and including extinct species
igtree.extant is the tree in simmap format with incipient and good flags without extinct species
recontree is the reconstructed tree in phylo format, reconstructed using the approximation in Lambert et al. 2014
reconL is the matrix corresponding to recontree
L0 is a matrix where the crown age is at 0; for internal use only

See also

Author

Rampal S. Etienne

Examples

 pbd_sim(c(0.2,1,0.2,0.1,0.1),15)
#> $tree
#> 
#> Phylogenetic tree with 16 tips and 15 internal nodes.
#> 
#> Tip labels:
#>   S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> Rooted; includes branch length(s).
#> 
#> $stree_random
#> 
#> Phylogenetic tree with 13 tips and 12 internal nodes.
#> 
#> Tip labels:
#>   S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> Rooted; includes branch length(s).
#> 
#> $stree_oldest
#> 
#> Phylogenetic tree with 13 tips and 12 internal nodes.
#> 
#> Tip labels:
#>   S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> Rooted; includes branch length(s).
#> 
#> $stree_youngest
#> 
#> Phylogenetic tree with 13 tips and 12 internal nodes.
#> 
#> Tip labels:
#>   S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> Rooted; includes branch length(s).
#> 
#> $L
#>       [,1] [,2]       [,3]       [,4]       [,5] [,6]
#>  [1,]    1    0 15.0000000 15.0000000 -1.0000000    1
#>  [2,]    2    1  9.5201270  8.9613779 -1.0000000    2
#>  [3,]    3    2  7.3368944  6.6488420  5.8796394    3
#>  [4,]    4    1  6.4255775  6.2060125 -1.0000000    4
#>  [5,]    5    4  4.1076234  2.8235107 -1.0000000    6
#>  [6,]    6   -5  3.9199562  2.1044360 -1.0000000    7
#>  [7,]    7    1  3.4250797  0.4898537 -1.0000000   10
#>  [8,]    8   -6  2.8801156  2.8358924 -1.0000000    5
#>  [9,]    9    4  2.3308958  0.7490593 -1.0000000    9
#> [10,]   10   -6  2.1253450 -1.0000000  1.7678678    4
#> [11,]   11    2  1.6059138  0.9480349 -1.0000000    8
#> [12,]   12    2  0.8912839 -1.0000000 -1.0000000    2
#> [13,]   13    1 15.0000000 14.6029944  4.9758682   11
#> [14,]   14   13  5.8887757  4.1391231 -1.0000000   15
#> [15,]   15   13  5.6503684  4.8464600 -1.0000000   12
#> [16,]   16  -14  4.8602191  4.8102721  1.2755003   13
#> [17,]   17  -14  4.7451115  4.7373332 -1.0000000   14
#> [18,]   18   15  4.0141370  3.7521359 -1.0000000   16
#> [19,]   19   17  3.1875923  3.0311538  0.1255576   17
#> [20,]   20   14  2.3998997 -1.0000000 -1.0000000   15
#> [21,]   21   14  0.8441747 -1.0000000 -1.0000000   15
#> 
#> $sL_random
#>       [,1] [,2]       [,3]       [,4]       [,5] [,6]
#>  [1,]    1    0 15.0000000 15.0000000 -1.0000000    1
#>  [2,]    2    1  9.5201270  8.9613779 -1.0000000    2
#>  [3,]    3    2  7.3368944  6.6488420  5.8796394    3
#>  [4,]    4    1  6.4255775  6.2060125 -1.0000000    4
#>  [5,]    5    4  4.1076234  2.8235107 -1.0000000    6
#>  [6,]    6    5  3.9199562  2.1044360 -1.0000000    7
#>  [7,]    7    1  3.4250797  0.4898537 -1.0000000   10
#>  [8,]    8    6  2.8801156  2.8358924 -1.0000000    5
#>  [9,]    9    4  2.3308958  0.7490593 -1.0000000    9
#> [10,]   10    6  2.1253450 -1.0000000  1.7678678    4
#> [11,]   11    2  1.6059138  0.9480349 -1.0000000    8
#> [12,]   12    2  0.8912839 -1.0000000  0.0000000    2
#> [13,]   13    1 15.0000000 14.6029944  4.9758682   11
#> [14,]   14   13  5.8887757  4.1391231  0.0000000   15
#> [15,]   15   13  5.6503684  4.8464600 -1.0000000   12
#> [16,]   16   14  4.8602191  4.8102721  1.2755003   13
#> [17,]   17   14  4.7451115  4.7373332 -1.0000000   14
#> [18,]   18   15  4.0141370  3.7521359 -1.0000000   16
#> [19,]   19   17  3.1875923  3.0311538  0.1255576   17
#> [20,]   20   14  2.3998997 -1.0000000  0.0000000   15
#> [21,]   21   14  0.8441747 -1.0000000 -1.0000000   15
#> 
#> $sL_oldest
#>       [,1] [,2]       [,3]       [,4]       [,5] [,6]
#>  [1,]    1    0 15.0000000 15.0000000 -1.0000000    1
#>  [2,]    2    1  9.5201270  8.9613779 -1.0000000    2
#>  [3,]    3    2  7.3368944  6.6488420  5.8796394    3
#>  [4,]    4    1  6.4255775  6.2060125 -1.0000000    4
#>  [5,]    5    4  4.1076234  2.8235107 -1.0000000    6
#>  [6,]    6    5  3.9199562  2.1044360 -1.0000000    7
#>  [7,]    7    1  3.4250797  0.4898537 -1.0000000   10
#>  [8,]    8    6  2.8801156  2.8358924 -1.0000000    5
#>  [9,]    9    4  2.3308958  0.7490593 -1.0000000    9
#> [10,]   10    6  2.1253450 -1.0000000  1.7678678    4
#> [11,]   11    2  1.6059138  0.9480349 -1.0000000    8
#> [12,]   12    2  0.8912839 -1.0000000  0.0000000    2
#> [13,]   13    1 15.0000000 14.6029944  4.9758682   11
#> [14,]   14   13  5.8887757  4.1391231 -1.0000000   15
#> [15,]   15   13  5.6503684  4.8464600 -1.0000000   12
#> [16,]   16   14  4.8602191  4.8102721  1.2755003   13
#> [17,]   17   14  4.7451115  4.7373332 -1.0000000   14
#> [18,]   18   15  4.0141370  3.7521359 -1.0000000   16
#> [19,]   19   17  3.1875923  3.0311538  0.1255576   17
#> [20,]   20   14  2.3998997 -1.0000000  0.0000000   15
#> [21,]   21   14  0.8441747 -1.0000000  0.0000000   15
#> 
#> $sL_youngest
#>       [,1] [,2]       [,3]       [,4]       [,5] [,6]
#>  [1,]    1    0 15.0000000 15.0000000 -1.0000000    1
#>  [2,]    2    1  9.5201270  8.9613779  0.0000000    2
#>  [3,]    3    2  7.3368944  6.6488420  5.8796394    3
#>  [4,]    4    1  6.4255775  6.2060125 -1.0000000    4
#>  [5,]    5    4  4.1076234  2.8235107 -1.0000000    6
#>  [6,]    6    5  3.9199562  2.1044360 -1.0000000    7
#>  [7,]    7    1  3.4250797  0.4898537 -1.0000000   10
#>  [8,]    8    6  2.8801156  2.8358924 -1.0000000    5
#>  [9,]    9    4  2.3308958  0.7490593 -1.0000000    9
#> [10,]   10    6  2.1253450 -1.0000000  1.7678678    4
#> [11,]   11    2  1.6059138  0.9480349 -1.0000000    8
#> [12,]   12    2  0.8912839 -1.0000000 -1.0000000    2
#> [13,]   13    1 15.0000000 14.6029944  4.9758682   11
#> [14,]   14   13  5.8887757  4.1391231  0.0000000   15
#> [15,]   15   13  5.6503684  4.8464600 -1.0000000   12
#> [16,]   16   14  4.8602191  4.8102721  1.2755003   13
#> [17,]   17   14  4.7451115  4.7373332 -1.0000000   14
#> [18,]   18   15  4.0141370  3.7521359 -1.0000000   16
#> [19,]   19   17  3.1875923  3.0311538  0.1255576   17
#> [20,]   20   14  2.3998997 -1.0000000  0.0000000   15
#> [21,]   21   14  0.8441747 -1.0000000 -1.0000000   15
#> 
#> $igtree.extinct
#> 
#> Phylogenetic tree with 21 tips and 20 internal nodes.
#> 
#> Tip labels:
#> 	S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> The tree includes a mapped, 2-state discrete character
#> with states:
#> 	g, i
#> 
#> Rooted; includes branch lengths.
#> 
#> $igtree.extant
#> 
#> Phylogenetic tree with 16 tips and 15 internal nodes.
#> 
#> Tip labels:
#> 	S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> The tree includes a mapped, 2-state discrete character
#> with states:
#> 	g, i
#> 
#> Rooted; includes branch lengths.
#> 
#> $recontree
#> 
#> Phylogenetic tree with 13 tips and 12 internal nodes.
#> 
#> Tip labels:
#>   S1-1-1, S10-10-7, S4-4-4, S9-9-9, S6-6-5, S7-7-6, ...
#> 
#> Rooted; includes branch length(s).
#> 
#> $reconL
#>       [,1] [,2]      [,3]      [,4] [,5] [,6]
#>  [1,]    1    0 15.000000 15.000000   -1    1
#>  [2,]    2    1  9.520127  9.520127   -1    2
#>  [3,]    4    1  6.425578  6.425578   -1    4
#>  [4,]    5    4  4.107623  4.107623   -1    6
#>  [5,]    6    5  3.919956  3.919956   -1    7
#>  [6,]    7    1  3.425080  3.425080   -1   10
#>  [7,]    8    6  2.880116  2.880116   -1    5
#>  [8,]    9    4  2.330896  2.330896   -1    9
#>  [9,]   11    2  1.605914  1.605914   -1    8
#> [10,]   14   15  5.888776  5.888776   -1   15
#> [11,]   15    1 15.000000 15.000000   -1   12
#> [12,]   17   14  4.745112  4.745112   -1   14
#> [13,]   18   15  4.014137  4.014137   -1   16
#> 
#> $L0
#>       [,1] [,2]          [,3]       [,4]      [,5] [,6]
#>  [1,]    1    0 -0.0000000001  0.0000000 -1.000000    1
#>  [2,]    2    1  5.4798729777  6.0386221 -1.000000    2
#>  [3,]    3    2  7.6631055649  8.3511580  9.120361    3
#>  [4,]    4    1  8.5744224870  8.7939875 -1.000000    4
#>  [5,]    5    4 10.8923766299 12.1764893 -1.000000    6
#>  [6,]    6   -5 11.0800437551 12.8955640 -1.000000    7
#>  [7,]    7    1 11.5749202664 14.5101463 -1.000000   10
#>  [8,]    8   -6 12.1198844482 12.1641076 -1.000000    5
#>  [9,]    9    4 12.6691042361 14.2509407 -1.000000    9
#> [10,]   10   -6 12.8746550464 -1.0000000 13.232132    4
#> [11,]   11    2 13.3940861570 14.0519651 -1.000000    8
#> [12,]   12    2 14.1087160901 -1.0000000 -1.000000    2
#> [13,]   13    1  0.0000000000  0.3970056 10.024132   11
#> [14,]   14   13  9.1112242605 10.8608769 -1.000000   15
#> [15,]   15   13  9.3496315706 10.1535400 -1.000000   12
#> [16,]   16  -14 10.1397809088 10.1897279 13.724500   13
#> [17,]   17  -14 10.2548884790 10.2626668 -1.000000   14
#> [18,]   18   15 10.9858630020 11.2478641 -1.000000   16
#> [19,]   19   17 11.8124076606 11.9688462 14.874442   17
#> [20,]   20   14 12.6001003282 -1.0000000 -1.000000   15
#> [21,]   21   14 14.1558253397 -1.0000000 -1.000000   15
#>